# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kivo Hub API",
    "version": "1.0.0",
    "description": "Backend API for the Gaming Center Management System."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["name", "gmail", "phone_number", "password"],
        "properties": {
          "name": { "type": "string", "example": "John Doe" },
          "gmail": { "type": "string", "example": "user1@gmail.com" },
          "phone_number": { "type": "string", "example": "+998123456789" },
          "password": { "type": "string", "example": "user123" }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["phone_number", "password"],
        "properties": {
          "phone_number": { "type": "string", "example": "+998123456789" },
          "password": { "type": "string", "example": "user123" }
        }
      },
      "BookingCalculate": {
        "type": "object",
        "required": ["tariff_id", "asset_ids", "start_time", "end_time"],
        "properties": {
          "tariff_id": { "type": "integer", "example": 1 },
          "asset_ids": { "type": "array", "items": { "type": "integer" }, "example": [1, 2] },
          "start_time": { "type": "string", "format": "date-time", "example": "2026-04-24 10:00:00" },
          "end_time": { "type": "string", "format": "date-time", "example": "2026-04-24 12:00:00" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRegister" }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLogin" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/bookings/calculate": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Calculate cost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookingCalculate" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/trades": {
      "get": {
        "tags": ["Finance"],
        "summary": "Fetch ledger",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}